add/ sistema de muerte, estado de muerte y explosion

+++ include/Popup.h
@@ -9,8 +9,8 @@
 class Popup
 {
     public:
-        Popup(string ruta, int life);
-        void setLife(int t);
+        Popup(string ruta, double life);
+        void setLife(double t);
         SuperSprite* getSprite();
         void setPosition(double x, double y);
         void drawPopup(sf::RenderWindow *w, double i);
@@ -25,7 +25,7 @@ class Popup
     private:
     SuperSprite* popup;
     sf::Clock popup_clock;
-    int popup_life;
+    double popup_life;
     bool alive;
     vector <double> pos;
 };
